refactor(app): 拆分窗口对象图与来源视图
